air-trust/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "air-trust"
version = "0.4.0"
description = "Universal compliance trust layer for AI systems. One install, any framework. HMAC-SHA256 audit chain, agent identity, policy enforcement, CSA Agentic Trust Framework (ATF) conformance, verify CLI."
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.9"
authors = [
    { name = "Jason Shotwell", email = "jason@airblackbox.ai" },
]
keywords = [
    "ai", "compliance", "audit", "trust", "eu-ai-act",
    "langchain", "openai", "anthropic", "crewai", "llm",
    "safety", "governance", "pii", "injection-detection",
    "agent-identity", "article-14", "permissions",
    "policy-enforcement", "verify", "audit-export",
    "atf", "agentic-trust-framework", "csa", "zero-trust",
    "agent-governance", "agent-urn", "iso-42001",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Security",
]

# Zero required dependencies — pure Python
dependencies = []

[project.optional-dependencies]
# Install extras for framework-specific features
langchain = ["langchain-core>=0.1.0"]
llamaindex = ["llama-index-core>=0.10.0"]
haystack = ["haystack-ai>=2.0.0"]
openai = ["openai>=1.0.0"]
anthropic = ["anthropic>=0.20.0"]
google = ["google-generativeai>=0.5.0"]
crewai = ["crewai>=0.30.0"]
smolagents = ["smolagents>=0.1.0"]
pydantic-ai = ["pydantic-ai>=0.1.0"]
dspy = ["dspy>=2.0.0"]
otel = ["opentelemetry-sdk>=1.20.0"]
mcp = ["mcp>=1.0.0"]

# All frameworks
all = [
    "langchain-core>=0.1.0",
    "openai>=1.0.0",
    "anthropic>=0.20.0",
]

# Development
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

[project.urls]
Homepage = "https://airblackbox.ai"
Repository = "https://github.com/airblackbox/air-trust"
Documentation = "https://airblackbox.ai/docs"
Issues = "https://github.com/airblackbox/air-trust/issues"

[tool.hatch.build.targets.wheel]
packages = ["air_trust"]

[tool.ruff]
target-version = "py39"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
